Extract risk management domain models
